requirements.txt
azure-functions
azure-functions-durable
azure-storage-blob
azure-data-tables
flask
requests